Refactor DecksNew
Extract DeckList, CardsCatalogue sections into own components
Fix small visual bugs
